requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
accelerate>=0.20.0
diffusers>=0.21.0
transformers>=4.30.0
peft>=0.5.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
xformers>=0.0.20